LikeStarterDjango/acmeconf/requirements.txt
django_booking==0.7.2
Django==2.0.3
drf_extensions==0.3.1
zeep==2.5.0
djangorestframework==3.8.2
